refactor out DraftPostScheduler
